+++ simtools/__init__.py
@@ -10,7 +10,8 @@ __author__ = "Przemyslaw (Mack) Nowak"
 
 from .argparse import parse_args
 from .params import load_params, Params
+from .random import generate_seed
 from .simrun import (generate_sim_dirname, generate_sim_id, make_dirs,
                      norm_executable, run_sim)
 from .utils import save_versions
-from . import argparse, params, simrun, utils
+from . import argparse, params, random, simrun, utils